% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pactaverse.R
\docType{package}
\name{pactaverse-package}
\alias{pactaverse}
\alias{pactaverse-package}
\title{pactaverse: Easily Install and Load the 'PACTA-verse'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'pactaverse' is a set of packages that work to help R users implement various functionality related to the PACTA open source project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmi-pacta.github.io/pactaverse/}
  \item Report bugs at \url{https://github.com/RMI-PACTA/pactaverse/issues}
}

}
\author{
\strong{Maintainer}: Jackson Hoffart \email{jackson.hoffart@gmail.com}

Other contributors:
\itemize{
  \item RMI [copyright holder, funder]
}

}
\keyword{internal}
