% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{pactaverse_conflicts}
\alias{pactaverse_conflicts}
\title{Conflicts between the pactaverse and other packages}
\usage{
pactaverse_conflicts(only = NULL)
}
\arguments{
\item{only}{Set this to a character vector to restrict to conflicts only
with these packages.}
}
\description{
This function lists all the conflicts between packages in the pactaverse
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
pactaverse_conflicts()
}
