% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{pactaverse_deps}
\alias{pactaverse_deps}
\title{List all pactaverse dependencies}
\usage{
pactaverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
pactaverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all pactaverse dependencies
}
