% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{pactaverse_logo}
\alias{pactaverse_logo}
\title{The pactaverse logo, using ASCII or Unicode characters}
\usage{
pactaverse_logo(unicode = cli::is_utf8_output())
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\description{
Use \code{\link[cli:ansi_strip]{cli::ansi_strip()}} to get rid of the colors.
}
\examples{
pactaverse_logo()
}
