% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pactaverse_packages}
\alias{pactaverse_packages}
\title{List all packages in the pactaverse}
\usage{
pactaverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include pactaverse in the list?}
}
\description{
List all packages in the pactaverse
}
\examples{
pactaverse_packages()
}
