% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pactaverse.R
\docType{package}
\name{pactaverse-package}
\alias{pactaverse}
\alias{pactaverse-package}
\title{pactaverse: Easily Install and Load the 'PACTA-verse'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'pactaverse' is a set of packages that work to help R users implement various functionality related to the PACTA open source project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmi-pacta.github.io/pactaverse/}
  \item Report bugs at \url{https://github.com/RMI-PACTA/pactaverse/issues}
}

}
\author{
\strong{Maintainer}: Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID}) [contractor]

Authors:
\itemize{
  \item CJ Yetman \email{cj@cjyetman.com} (\href{https://orcid.org/0000-0001-5099-9500}{ORCID}) [contractor]
  \item Jacob Kastl \email{jacob.kastl@gmail.com} [contractor]
  \item Monika Furdyna \email{monika.furdyna@gmail.com} [contractor]
  \item Alex Axthelm \email{aaxthelm@rmi.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item RMI \email{PACTA4investors@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
